+++ CHANGELOG.md
@@ -32,6 +32,9 @@ Current
 
 ### Changed:
 
+- [No match found due to schema mismatch now a `500 Internal Server Error` response instead of a `400 Bad Request` response](https://github.com/yahoo/fili/pull/113)
+    * This should never be a user fault, since that check is much earlier
+
 - [Make `SegmentMetadata::equals` `null`-safe](https://github.com/yahoo/fili/pull/113)
     * It was not properly checking for `null` before and could have exploded